move the file chooser implementation to WebxdcActivity class
