+++ src/org/thoughtcrime/securesms/WebxdcActivity.java
@@ -16,12 +16,15 @@ import android.view.Menu;
 import android.view.MenuItem;
 import android.webkit.JavascriptInterface;
 import android.webkit.MimeTypeMap;
+import android.webkit.ValueCallback;
+import android.webkit.WebChromeClient;
 import android.webkit.WebResourceResponse;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 import androidx.core.app.TaskStackBuilder;
 import androidx.core.content.pm.ShortcutInfoCompat;
 import androidx.core.content.pm.ShortcutManagerCompat;
@@ -49,6 +52,8 @@ import java.util.Map;
 
 public class WebxdcActivity extends WebViewActivity implements DcEventCenter.DcEventDelegate  {
   private static final String TAG = WebxdcActivity.class.getSimpleName();
+  private static final int REQUEST_CODE_FILE_PICKER = 51426;
+  private ValueCallback<Uri[]> filePathCallback;
   private DcContext dcContext;
   private DcMsg dcAppMsg;
   private String baseURL;
@@ -93,6 +98,21 @@ public class WebxdcActivity extends WebViewActivity implements DcEventCenter.DcE
   @Override
   protected void onCreate(Bundle state, boolean ready) {
     super.onCreate(state, ready);
+
+    webView.setWebChromeClient(new WebChromeClient() {
+      @Override
+      @RequiresApi(21)
+      public boolean onShowFileChooser(WebView webView, ValueCallback<Uri[]> filePathCallback, WebChromeClient.FileChooserParams fileChooserParams) {
+        WebxdcActivity.this.filePathCallback = filePathCallback;
+        Intent intent = new Intent(Intent.ACTION_GET_CONTENT);
+        intent.addCategory(Intent.CATEGORY_OPENABLE);
+        intent.setType("*/*");
+        intent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, fileChooserParams.getMode() == FileChooserParams.MODE_OPEN_MULTIPLE);
+        WebxdcActivity.this.startActivityForResult(Intent.createChooser(intent, getString(R.string.select)), REQUEST_CODE_FILE_PICKER);
+        return true;
+      }
+    });
+
     DcEventCenter eventCenter = DcHelper.getEventCenter(WebxdcActivity.this.getApplicationContext());
     eventCenter.addObserver(DcContext.DC_EVENT_WEBXDC_STATUS_UPDATE, this);
     
@@ -301,6 +321,31 @@ public class WebxdcActivity extends WebViewActivity implements DcEventCenter.DcE
     } catch(Exception e) {
       Toast.makeText(context, "ErrAddToHomescreen: " + e, Toast.LENGTH_LONG).show();
     }
+  }
+
+  @Override
+  public void onActivityResult(int reqCode, int resultCode, final Intent data) {
+    if (reqCode == REQUEST_CODE_FILE_PICKER && filePathCallback != null) {
+      Uri[] dataUris = null;
+      if (resultCode == Activity.RESULT_OK && data != null) {
+        try {
+          if (data.getDataString() != null) {
+            dataUris = new Uri[]{Uri.parse(data.getDataString())};
+          } else if (data.getClipData() != null) {
+            final int numSelectedFiles = data.getClipData().getItemCount();
+            dataUris = new Uri[numSelectedFiles];
+            for (int i = 0; i < numSelectedFiles; i++) {
+              dataUris[i] = data.getClipData().getItemAt(i).getUri();
+            }
+          }
+        } catch (Exception e) {
+          e.printStackTrace();
+        }
+      }
+      filePathCallback.onReceiveValue(dataUris);
+      filePathCallback = null;
+    }
+    super.onActivityResult(reqCode, resultCode, data);
   }
 
   class InternalJSApi {
